Update patch set 2
Patch Set 2:

Just realized the reason I kept the "top comment lines about automatically generated, do not edit.". The reason is, if some config changed, it will be convenient to copy the config file back from kernel/.config,and then git review to upload the changes. If we remove that comments, it means we will need to manually update this config file and that's not conveient.

Your idea?

Patch-set: 2
Label: Verified=0