Update patch set 4
Patch Set 4: Code-Review-1

Since this is just a new file in an existing directory then sure, this is OK.  It'll be interesting to see if we can push this upstream but that's a problem for another day.

Note that none of the current config files in that directory have the `auto generated' comment lines so I'm going back to a negative vote based on that.

Patch-set: 4
Label: Code-Review=-1