Add the "timers: do not raise softirq unconditionally" temply
This patch enable the nohz_full and is important for RT. Bring it back
temply, while waiting for more work on RT community.

Please refer to https://lkml.org/lkml/2015/3/17/783 for more information
of the revert.

A little rebase needed because it's reverted on old code base.

Upstream status: pending

Change-Id: I2747e087faf4145b69b800a60b8d9414bc71e206