Update patch set 2
Patch Set 2: Code-Review-1

(4 comments)

A few issues inline, 2 of which are serious.

This is more a sylistic issue so I wouldn't -1 on it but you really don't need to put {} around every single shell variable.  A simple variable that is surrounded by white space is much more readable without the {}.

Patch-set: 2
Label: Code-Review=-1