Add verify scripts for continuous integration
The scripts require a host system running CentOS 7.1 and a guest image
also running CentOS 7.1 (the image will be uploaded to Artifactory).

The verify script needs to connect to the host as root through ssh without
providing a password.  This script should be run from the root of kvmfornfv
source tree.

The README file contains more info about the host/guest-side scripts.

Change-Id: Ie933ea0089ac82acd39fc48088615215993312f3