Update patch set 7
Patch Set 7:

It seems like it's taking a configuration file I wasn't aware of, from the kernel tree. What is the value of an specific config given the fact that we don't know which distro is tested in?

Also, shouldn't it be better to pass the locations as parameters? That way the CI environment has more flexibility.

Patch-set: 7
Reviewer: Gerrit User 1085 <1085@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0