Update patch set 7
Patch Set 7:

Sure! I was meaning to make usage of the oldconfig target in the kernel Makefile(that makes the configuration depend in the previous kernel config, if any), and change the values that you need to change. For instance, as the RT wiki specify, there is the PREMPT_RT option, select the deadline scheduler and drop, if possible, the acpi interface in the kernel.

That way, the configuration would be more respectful with the distro used, and you'll enjoy the RT features.

Patch-set: 7