Update patch set 7
Patch Set 7:

The purpose of this script is to build the kernel automatically in the CI environment without any user intervention.  oldconfig will prompt the user for config items not in the existing config file.  The user has to decide how to set them.  I do not think this can or should be automated.

The current approach is that one runs make oldconfig making the config decisions manually and carefully, then copies the resulting .config to opnfv.config and uploads it to gerrit, so others can review and verify it.

Patch-set: 7