[WIP] avoid needless data processing
By enabling the log dirty before vm starting, we can skip the processing
of the VM's unused pages in the ram bulk stage. This patch can reduce the
total live migraion time about 60% for an idle 8G RAM guest.

Please do not apply this patch because it will impact the performance,
it's just for protype. A pv solution will be used instead.

Change-Id: I434866c377b37d69f03c329f857e18abcb596136
Signed-off-by: Liang Li <liang.z.li@intel.com>