Add the clocksource_watchdog kernel config
The CLOCKSOURCE_WATCHDOG is always added automatically whenever a new
kernel configuration created.

The reason is because when we created the original kernel config file, we
disabled the clock source watchdog by force. Please refer to
https://gerrit.opnfv.org/gerrit/#/c/1293/ for more information. Since now that patch is not removed, we should add the CLOCKSOURCE_WATCHDOG back to our kernel config file.

As stated on the abandon reason of
https://gerrit.opnfv.org/gerrit/#/c/1293/, we should use tsc=reliable kernel
parameter to achieve the purpose. Also we are preparing another patch to
upstream linux to detect if the TSC is reliable automatically.

Change-Id: I89a78d5899c0330776bb167331a1c6750fc0f50a
Signed-off-by: Yunhong Jiang <yunhong.jiang@intel.com>