Add configuration to support OVS kernel module
OVS kernel module is important for NFV. The existed kernel config misses
several important kernel configuration, mostly for VxLAN, which is required
by the OVS kernel module.

Also add the OVS as a module, so that user can use it if needed, or replace
it when use the OPNFV patched OVS module.

Change-Id: I032f84e0468fc2614557274a50f30d502d39cc52
Signed-off-by: Yunhong Jiang <yunhong.jiang@intel.com>