Update patch set 1
Patch Set 1:

Jose, so you mean it's assumed as kernel module, instead of built-in? That's a bit strange :)

Not sure why this patch failed to be merged, have a quick check on https://build.opnfv.org/ci/job/kvmfornfv-verify/59/console and possibly a transient error. Let's try it again.

Patch-set: 1