Update patch set 1
Patch Set 1: Code-Review-1

We need to provide some measured results that show there is a benefit before we make a change like this.  Note, the -1 is not a reflection on the patch, I'd be happy to change this if you can show a performance benefit.

Patch-set: 1
Reviewer: Gerrit User 1453 <1453@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1