Update patch set 4
Patch Set 4:

Can you clarify your intentions with this patch ? It isn't clear from your commit log. 

I think that you are saying that you are temporarily restoring this change, that
was reverted in the upstream preempt-rt.

If that is the case, can you elaborate on what advantages this brings, and how if
it is restored do we avoid the problems the threads talk about ? i.e. how are we
avoiding BUG_ON ?

Patch-set: 4
Label: Verified=0