Update patch set 4
Patch Set 4:

> Hi, Bruce,
 > Thanks for your review. I think the major benefit of this patch is
 > to add the NOHZ_FULL support back to the RT kernel, as stated on
 > the commit message. It does not fix the BUG_ON issue, so is just a
 > dirty patch. There has been discussion on the RT side on the
 > try_lock.

I've run KVM -rt with, and without full_nohz enabled. So if there are
issues with running with full_nohz, I can probably help.

That being said, standard NOHZ shows nearly as well in many benchmarks,
since there is overhead introduced by full_nohz, and since we aren't
to the point of complete isolation, the busy vcpu threads are being
interrupted already.

my caution is to benchmark with and without it enabled, since the
stability and less overhead are of value as well .. or at least, make
it an optional config.

Bruce

 > 
 > If there is no progress on the RT upstream, I think we have to fix
 > the issue discussed on the original thread ourselves and push to RT
 > upstream.
 > 
 > Thanks
 > --jyh

Patch-set: 4