Add patch to Linux 4.4 to support OVS
This file is a temporary patch that enables Linux 4.4 support in OVS.
Unfortunately, the OVS maintainers believe that the Linux 4.4 support
should be added in a different way so this patch has been rejected
but this is the only way to add 4.4 support for now.  We will only keep
this patch until the OVS source tree has officially added support for
Linux 4.4.

Upstream status: NA

Signed-off-by: Don Dugger <n0ano@n0ano.com>