Add patch to Linux 4.4 to support OVS
Given that OVS doesn't support Linux 4.4 yet we need to add a patch
to the OVS sources so that we can build a 4.4 version of the OVS
loadable kernel module that works with the RT Linux 4.4 kernel used
in OPNFV.

The directory `patches/ovs' contains patches (currently only one) that
are applied against the OVS tree.  The Fuel build script is modified
to apply all of the patches in this directory to the OVS tree.  Then
a working OVS KLM is created that is then inserted into the RT kernel
DEB package so that the end result is an RT kernel that supports OVS.

Upstream status: NA

Change-Id: I361f92526fb4bcafbeab9ce21570202f4aad1632
Signed-off-by: Don Dugger <n0ano@n0ano.com>