Update patch set 1
Patch Set 1:

> My understanding is that the OVS4NFV project just uses the upstream
 > OVS tree, they have no plans to create their own changes against
 > that tree.

Thanks for the clarification. So I'm a bit lost. Can you please give more hints on how this patch will be used? For example, which binary will this patch be included to build for? Per my understanding, it's not for either kernel or qemu. Instead, it's for the OVS and we don't build any OVS package. So you mean this is only for reference, so that if the end user want to use our 4.4 tree, they can fetch this patch, apply to the OVS tree and apply it? 

Sorry for the confusion.
 > 
 > The upstream OVS community is working on adding Linux 4.4 support
 > but they don't seem in any great hurry and I don't know of a
 > specific target date, hence we can use this patch for now and throw
 > it away if OVS magically adds 4.4 support before the Colorado
 > release (I check the OVS tree daily on this isue).
 > 
 > Not sure why we would want to chance the OVS user space tools, this
 > is just internal kernel interfaces, there shouldn't be a user space
 > impact.

 > My understanding is that the OVS4NFV project just uses the upstream
 > OVS tree, they have no plans to create their own changes against
 > that tree.
 > 
 > The upstream OVS community is working on adding Linux 4.4 support
 > but they don't seem in any great hurry and I don't know of a
 > specific target date, hence we can use this patch for now and throw
 > it away if OVS magically adds 4.4 support before the Colorado
 > release (I check the OVS tree daily on this isue).
 > 
 > Not sure why we would want to chance the OVS user space tools, this
 > is just internal kernel interfaces, there shouldn't be a user space
 > impact.

Patch-set: 1
Reviewer: Gerrit User 473 <473@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0