Move quirks out of build script
Due to expediencyh concerns some quirk code was placed in the build
script in order to get the Brahmaputra release out on time.  These
quirks don't belong in the script itself and, now that there is a
mechanisms to applly arbitrary patch files during the build script,
this patch moves that quirk code out to the separate patch file
area.

Upstream status - NA

Change-Id: Ib8100eef00009bbaf0e16b28849821ef5878f9f5
Signed-off-by: Donald Dugger <n0ano@n0ano.com>