Update patch set 1
Patch Set 1:

(7 comments)

I obviously lack the context about this patch but there are many problems with patch as commented on the files.

Even though you allow this change to go in, corresponding change in releng repo will not go in its current state since many things come from local directories and we do not have possibility of reproducing stuff. And the patch lacks the basic explanation regarding what is expected to be achieved.

https://gerrit.opnfv.org/gerrit/#/c/16035/

Patch-set: 1
Label: Verified=0