Reduce the kernel package size
Currently the kernel rpm package and the kernel-devel rpm package are
both 150M+. These rpm packages will be uploaded to the artifact repo and
downloaded on each test, and will be kept on the artifact repo for
several days. Such big package is overhead for both network and storage.

The reasons of this big size are:
a) The kernel modules are not stripped.
b) A vmlinux.bz2 is contained on the kernel rpm, which is about 90M.
c) The kernel-devel package includes all the source code files.

This patchset try to avoid the overhead from item a/b.

a) A new parameter, "-d", is added to the ci/build.sh. If this option is
not specified, the kernel module will be stripped.
b) We don't copy the vmlinux.bz2 anymore. This is not so good in long
term. We should copy the vmlinux.bz2 seperatedly when "-d" parameter is
specified. It's not easy to achieve and need more investigation. But at
should be ok for our situation, but should be fixed before the final
release.

With above two changes, the kernel package will be 15M only, but the
kernel-devel package is still too big is next step.

In the long term, we should follow below rules:
a) On the patch verification/merge task, we should not use '-d'
parameter. The stripped package will be passed back to releng.

b) On release building or building on the stable branch, we should use
'-d' parameter to get all the debug information, since that will be
useful for debugging.

c) On daily building, I'm not sure if we need the debug info and it's
open for discussion.

d) We should discuss with upstream kernel community on how to handle
this. I think such big rpm package is not good even for normal kernel
developer if they prefer rpm method. I have a discussion with one author
of script/package/mkspec, and he is open for this change.

upstream-status: N/A

Change-Id: Ie93e7b179d8530d5e7eee04a23a48d4a6a6ee69e
Signed-off-by: Yunhong Jiang <yunhong.jiang@linux.intel.com>