Reduce the kernel package size
Currently there kernel packet built out is about 200M, which is too big
and may cost time for uploading and downloading.

The reasons of this big size are:
a) The kernel modules are not stripped.
b) A vmlinux.bz2 is contained on the kernel rpm, which is about 90M.

Some quick change to avoid above reasons. This is not a good solution in
the long term, the final solution should be something like the
distribution, with kernel packages and debug info packages. For generic
jjb job like patch verification and patch merge, we don't need the debug
packages, but only the stripped version. For release building, we need
everything, or at least we need keep everything, although only the
stripped kernel packages are included on the ISO file.

Current upstream kernel does not support debug info packages yet.
We need a clean solution for it in the long term.

Change-Id: Ie93e7b179d8530d5e7eee04a23a48d4a6a6ee69e
Signed-off-by: Yunhong Jiang <yunhong.jiang@linux.intel.com>