Update patch set 3
Patch Set 3:

I think the qemu-kvm is not a built out from the source tree. It should be something like qemu-system-x86_64 which is normally installed on /usr/local/bin/.

BTW, how many parrellel threads are used on the building,i.e. the value of make -j? It's not specified explicitly, so it should be the default value of rpmbuild, right? Do you know the default value of the rpmbuild?

Thanks
--jyh

Patch-set: 3
Label: Verified=0