Update patch set 1
Patch Set 1:

The tc/measurement name depends on the name of the .yaml file. If we have same yaml file name for all scenarios, the measurement name in influxdb will be same and the results of all the tests will be stored in that single measurement. It will be easy to differentiate the results if we have different names for each scenario.Please correct me if I am missing something here.

Patch-set: 1
Reviewer: Gerrit User 5447 <5447@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0