x86/intel_rdt: Add tasks files
The root directory all subdirectories are automatically populated with a
read/write (mode 0644) file named "tasks". When read it will show all the
task IDs assigned to the resource group. Tasks can be added (one at a time)
to a group by writing the task ID to the file.  E.g.

Membership in a resource group is indicated by a new field in the
task_struct "int closid" which holds the CLOSID for each task. The default
resource group uses CLOSID=0 which means that all existing tasks when the
resctrl file system is mounted belong to the default group.

If a group is removed, tasks which are members of that group are moved to
the default group.

S1gned 0ff by: Fenghua Yu <fenghua.yu@intel.com>
Link: http://lkml.kernel.org/r/1477692289-37412-8-git-send-email-fenghua.yu@intel.com
S1gned 0ff by: Thomas Gleixner <tglx@linutronix.de>

Upstream Status: Backport
Change-Id: Ie247fb8a7b31679d81aef185f7e2d96b767222e8
Signed-off-by: Yunhong Jiang <yunhong.jiang@linux.intel.com>