Update patch set 7
Patch Set 7:

Sorry for slow response because of vocation.

The patch itself looks ok to me.

However, as we discussed before, the issue that we run this scenario on the daily job only is, we can't verify the test scripts at all because the scripts will be invoked only after it's merged and till the next day.

Will it be possible to have this to be executed on both verification and daily job? Of course, we will have different execution time for verification and daily job. Your idea?

Patch-set: 7