Update patch set 4
Patch Set 4:

(3 comments)

Thanks for the patch.

I have some comments inline. The big question is, are there any difference of the new code and the old code? It's not easy to compare them manually because of the big changes. At lease I noticed the "VERSION" change, which seems is not on the old code. Anymore?

Patch-set: 4