Update patch set 2
Patch Set 2:

Thanks for the update.

I also considered that if we should restore back the original one after the testing, and then realized that we don't need restore back if we will use our qemu for all the testing.

Will wait till the PoD is back, so that we can make sure it works.

Patch-set: 2