Update patch set 1
Patch Set 1:

Hi Yunhong, thanks for the comment.

The issue is not with built out is too big.
Earlier  ericsson-build3 jenkins slave is used for kvmfornfv build job. Now it is ericsson-build4. Earlier builds for this patch failed on ericsson-build4 because of some configuration missing on that slave. I have informed the same to fdegir, he did some configuration that was missing (may be gsutil, not sure). Now the build, uploading and our verification job is working fine.

Patch-set: 1
Label: Verified=0