Update patch set 8
Patch Set 8:

(1 comment)

Hi Yunhong, thanks for the comments.

Before optimization:  For every testcase, we are doing docker image build and running a container. docker image build(yardstick installation etc.,) takes 3-4 mins.so for all the testcases,it takes 12-16 min for this build process.  

After optimization :  In this patch, docker image build is done only once,so reducing approximately 9-12 min of time and we are using -d deamon option (run -i -d -v) so that, the same container will be used for all the cyclictest testcases.

Patch-set: 8