APIC virtualization fix and cleanup
This is back ported from Paolo Bonzini's "cleanup and speedup for APICv"
patches.  See https://lkml.org/lkml/2016/12/19/352 for more details.

This fixes a performance bug where an interrupt from a device managed by
vfio-pci is queued in PIR after PIR has synchronized to VIRR.
This causes the interrupt not delivered to the guest until the next
VM exit-entry cycle.

Change-Id: I0383a3d5c4a533b5f05198466c604e801fa60e54
Signed-off-by: David Su <david.w.su@intel.com>