APIC virtualization fix and cleanup
This is back ported from Paolo Bonzini's "cleanup and speedup for APICv"
patches.  See https://lkml.org/lkml/2016/12/19/352 for more details.

This fixes a performance bug where an interrupt from a device managed by
vfio-pci is queued in PIR after PIR has synchronized to VIRR.
This causes the interrupt not delivered to the guest until the next
VM exit-entry cycle.

Change-Id: I3172f08c9fc4f82c5a178f1f902d4560356a9461
Signed-off-by: David Su <david.w.su@intel.com>