Disabling execution of cyclictest as part of verify job.
This patch is used to temporarily disable cyclictest execution as a part of
kvmfornfv verify job to speed up the verification process.

This is only a temporary workaround/solution as kvmfornfv verify job
is getting triggered for patches which includes only docs/ as well.

Note: Once complete documentation is made available for danube release
will revert this patch

Change-Id: Ib04a7d9481b74a0fafcd1d489b80c94051d875e6
Signed-off-by: Gundarapu Kalyan Reddy <reddyx.gundarapu@intel.com>