Update patch set 2
Patch Set 2:

Thanks for the information.

I have a quick check on the yardstick patch. A bit curios that why do we need https://gerrit.opnfv.org/gerrit/#/c/34457/2/yardstick/benchmark/scenarios/compute/disable_trace.sh there? Are there any special reason we want to put it to the yardstick? IMHO, yardstick is just a benchmark and should not care for the trace story etc. Will comment on that patch.

Patch-set: 2