Update patch set 2
Patch Set 2:

Hi, Jack and Kalyan, are there any special difference between the mechanism on this patch, i.e. the tear down environment script and the existed mechanism, i.e. the guest_setup_seqs etc? Is it possible that we provide the tear down scripts from outside?

The reason asking is, IMHO, the disable_trace scripts is not part of the benchmark, and also, what's the trace to be enabled/disabled is not part of the benchmark.

I'm not very well on yardstick, so please correct me if I'm wrong.

Thanks
--jyh

Patch-set: 2
Label: Code-Review=0