Update patch set 2
Patch Set 2:

Hi Yunhong,

The only difference between the mechanisms is the guest_setup_seqs  will use script_dir: "/opt/scripts" to copy and execute the configuration scripts on nodes where as node context teardown uses the path yardstick.benchmark.scenarios.compute. In both the cases the scripts should be local to yardstick so that it will upload and execute them on nodes mentioned. To make the scripts local in case of guest_setup_seqs we are copying the scripts to script_dir: "/opt/scripts" and in case of tear down we will copy to yardstick.benchmark.scenarios.compute at run time instead of making them a part of benchmark.

Patch-set: 2
Label: Verified=0