Generating only kernel debian packages
This patch includes scripts to generate only kernel debian packages
based on the commit-id provided in the compass conf file and also
used to remove code duplications.

NOTE: For generating the kernel debian packages commit-id should be
provided in compass.conf file and the default type should be changed
to 'ubuntu' in build.sh script.

Change-Id: I7f0822c565100a44e5289d6af825289e697130aa
Signed-off-by: Gundarapu Kalyan Reddy <reddyx.gundarapu@intel.com>